# deploy/helm/fuzekeys/files/openapi.yaml
...
openapi: 3.1.0
info:
  title: FuzeKeys API
  description: 'FuzeKeys — identity and credential vault. Managed digital identities, per-site accounts,
    encrypted vault assets, capability-token secret brokering, and automated signup workflows.


    DERIVED DOCUMENT: regenerate with `python scripts/export-openapi.py`. Do not hand-edit.'
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 2.0.0
paths:
...
  /api/v1/accounts/:
    get:
      tags:
      - Accounts
      summary: List Accounts
      description: List all accounts for the current user with stage information.
      operationId: get_accounts
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/AccountResponse'
                type: array
                title: Response List Accounts Api V1 Accounts  Get
      security:
      - HTTPBearer: []
    post:
      tags:
      - Accounts
      summary: Create Account
      description: Create a new account with stages.
      operationId: post_accounts
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/accounts/{account_id}/stages/{stage_id}:
    patch:
      tags:
      - Accounts
      summary: Update Account Stage
      description: Update a specific account stage.
      operationId: patch_accounts_by_account_id_stages_by_stage_id
      security:
      - HTTPBearer: []
      parameters:
      - name: account_id
        in: path
        required: true
        schema:
          type: integer
          title: Account Id
      - name: stage_id
        in: path
        required: true
        schema:
          type: integer
          title: Stage Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountStageUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/credentials/identity/{identity_id}/accounts:
    get:
      tags:
      - Credentials
      - Credentials
      summary: 👤 List Identity Accounts
      description: 'Get all accounts associated with a specific identity.


        Useful for understanding what accounts an identity has created and their status.'
      operationId: get_credentials_identity_by_identity_id_accounts
      parameters:
      - name: identity_id
        in: path
        required: true
        schema:
          type: integer
          description: Identity ID to get accounts for
          title: Identity Id
        description: Identity ID to get accounts for
      - name: x-api-key
        in: header
        required: true
        schema:
          type: string
          description: API key for service authentication
          title: X-Api-Key
        description: API key for service authentication
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/google/accounts/{identity_id}:
    get:
      tags:
      - Google Integration
      - Google Integration
      summary: Get Google Accounts
      description: 'Get all Google accounts for a specific identity.


        SECURITY (CRITICAL-1): requires auth; the identity must be owned by the

        caller (404 otherwise). The account lookup is additionally scoped through the

        owned identity so no cross-tenant account data is returned.'
      operationId: get_google_accounts_by_identity_id
      security:
      - HTTPBearer: []
      parameters:
      - name: identity_id
        in: path
        required: true
        schema:
          type: integer
          title: Identity Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Google Accounts Api Google Accounts  Identity Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AccountCreate:
      properties:
        website_name:
          type: string
          title: Website Name
        website_url:
          type: string
          title: Website Url
        identity_id:
          type: integer
          title: Identity Id
        website_domain:
          anyOf:
          - type: string
          - type: 'null'
          title: Website Domain
        account_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Account Type
          default: free
        signup_method:
          anyOf:
          - type: string
          - type: 'null'
          title: Signup Method
          default: automated
        stages:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Stages
      additionalProperties: false
      type: object
      required:
      - website_name
      - website_url
      - identity_id
      title: AccountCreate
...
    AccountResponse:
      properties:
        id:
          type: integer
          title: Id
        website_name:
          type: string
          title: Website Name
        website_url:
          type: string
          title: Website Url
        identity_id:
          type: integer
          title: Identity Id
        identity_name:
          type: string
          title: Identity Name
        is_active:
          type: boolean
          title: Is Active
        signup_completed:
          type: boolean
          title: Signup Completed
        created_at:
          type: string
          format: date-time
          title: Created At
        stages:
          items:
            $ref: '#/components/schemas/StageStatusResponse'
          type: array
          title: Stages
      type: object
      required:
      - id
      - website_name
      - website_url
      - identity_id
      - identity_name
      - is_active
      - signup_completed
      - created_at
      - stages
      title: AccountResponse
    AccountStageUpdate:
      properties:
        status:
          type: string
          title: Status
        error_message:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Message
        stage_data:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Stage Data
      type: object
      required:
      - status
      title: AccountStageUpdate
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    StageStatusResponse:
      properties:
        stage_type:
          type: string
          title: Stage Type
        stage_name:
          type: string
          title: Stage Name
        status:
          type: string
          title: Status
        attempts:
          type: integer
          title: Attempts
        started_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Started At
        completed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Completed At
        error_message:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Message
      type: object
      required:
      - stage_type
      - stage_name
      - status
      - attempts
      - started_at
      - completed_at
      - error_message
      title: StageStatusResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
...